move useClearPinButton hook
