+++ packages/react/src/components/controls/ClearPinButton.tsx
@@ -1,29 +1,8 @@
-import { setupClearPinButton } from '@livekit/components-core';
 import * as React from 'react';
-import { mergeProps } from '../../utils';
-import { useLayoutContext } from '../../context';
+import { useClearPinButton } from '../../hooks';
 
 /** @public */
 export interface ClearPinButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {}
-
-/** @public */
-export function useClearPinButton(props: ClearPinButtonProps) {
-  const { state, dispatch } = useLayoutContext().pin;
-
-  const buttonProps = React.useMemo(() => {
-    const { className } = setupClearPinButton();
-    const mergedProps = mergeProps(props, {
-      className,
-      disabled: !state?.length,
-      onClick: () => {
-        if (dispatch) dispatch({ msg: 'clear_pin' });
-      },
-    });
-    return mergedProps;
-  }, [props, dispatch, state]);
-
-  return { buttonProps };
-}
 
 /**
  * The ClearPinButton is a basic html button with the added ability to signal
